backend/aiServ/requirements.txt
flask
google-genai
scikit-learn
python-dotenv